+++ spec/unit/basic/operation_tree_builders/multiline_string_spec.rb
@@ -2,6 +2,41 @@
 
 require 'spec_helper'
 
+class WordOperationTreeBuilder < SuperDiff::Core::AbstractOperationTreeBuilder
+  def self.applies_to?(expected, actual)
+    expected.is_a?(::String) && actual.is_a?(::String)
+  end
+
+  protected
+
+  def unary_operations
+    expected_words = expected.split
+    actual_words = actual.split
+
+    Diff::LCS.diff(expected_words, actual_words).flat_map do |change_group|
+      change_group.map do |change|
+        SuperDiff::Core::UnaryOperation.new(
+          name: change.action == '+' ? :insert : :delete,
+          collection: expected,
+          key: change.position,
+          index: change.position,
+          value: change.element
+        )
+      end
+    end
+  end
+
+  def build_operation_tree
+    SuperDiff::Basic::OperationTrees::Array.new([])
+  end
+
+  private
+
+  def should_compare?(_operation, _next_operation)
+    false
+  end
+end
+
 RSpec.describe SuperDiff, type: :unit do
   describe '.diff' do
     subject(:diff) { SuperDiff.diff(expected, actual) }
@@ -109,6 +144,34 @@ RSpec.describe SuperDiff, type: :unit do
             plain_line '  It contains separate lines.\\n'
             expected_line '- What else can I say?\\n'
             actual_line '+ What else can I say?'
+          end
+            .to_s
+            .chomp
+        expect(diff).to eq(expected_output)
+      end
+    end
+
+    context 'when a string operation tree builder exists' do
+      around do |example|
+        with_configuration(extra_operation_tree_builder_classes: SuperDiff.configuration.extra_operation_tree_builder_classes + [WordOperationTreeBuilder]) do
+          example.run
+        end
+      end
+
+      let(:actual) { <<~STRING }
+        This here is a string.
+        It contains separate lines.
+        This one is different.
+      STRING
+
+      it 'does not attempt to diff each line' do
+        expected_output =
+          SuperDiff::Core::Helpers
+          .style(color_enabled: true) do
+            plain_line '  This here is a string.\\n'
+            plain_line '  It contains separate lines.\\n'
+            expected_line '- What else can I say?\\n'
+            actual_line '+ This one is different.\\n'
           end
             .to_s
             .chomp