+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Support elision for flat line trees. [#300](https://github.com/splitwise/super_diff/pull/300) by [@gschlager](https://github.com/gschlager)
 - Fix integration snippet indentation. [#299](https://github.com/splitwise/super_diff/pull/299) by [@gschlager](https://github.com/gschlager)
 - Pin all actions to full commit SHA. [#305](https://github.com/splitwise/super_diff/pull/305)
+- Do not attempt to sub-diff multiline strings. [#304](https://github.com/splitwise/super_diff/pull/304)
 
 ## 0.18.0 - 2025-12-05